Updated user registration
The user registration function now also uses the flexible httpRequest function to register users.

It also displays correct messages depending on if there are no values added to the inputs, the values are invalid or if they are valid and a user is registered.

If a user registration was successful, the user is logged in automatically and taken to the home page.