Updated profile page and cleanup
Profile page does now actually include the username, email, avatar and banner of the user with a neater layout.

Added a default banner image if user havent added it yet (WIP).

Users will be redirected to other pages if they try to access i.e. profile.html without being logged in.

nav-profile is now user-info and it's moved to the header (top left of pages).

General cleanup in the .scss department