Avatar and banner update pt1
Added ability to add avatar and banner image URLs during user registration. The inputs should have correct validation, so they will only accept publicly available URLs to images. I also added a button for each input to quickly clear the inputs, since URLs tend to be long. These inputs are optional and not required to finish user registration.

If a user adds URLs for avatar and/or banner, they will correctly show up as either their avatar across all pages and/or their profile banner on the profile page (depending on if they added URLs for only one of them or both).

Updated "message" function to only appear in ".message-container", since the plan is to always use that element. Also added a timer so when a message appears it dissapears after 3 seconds.

Added text and link to registration form and login form asking the user if they already have an account or need to create one, with links.

Other style changes across all pages