UrSpace v2
Redesigned the site to only have 3 pages, index, feed and profile (might be more later).

index.html is now the login and registration page. Users will have to log in or register a user to enter the site. This page has a new and cleaner layout with possibilities of both logging in and registering a new user.

feed.html is the new file name of the old index.html. No big changes to this page and profile.html.

Added new .svg logo for all pages with a slightly newer design.

Minor changes to file structure as well as removal of these now unused files:
- login.html
- signup.html
- login.js
- signup.js
- testlogin.html (only used when creating new layout)