Search done and new fetch/render post functions
Search functionality should work as intended now.

Made a separate function/file for fetching posts from the URL and one for rendering the posts as it got repeated quite a lot. Also made displaying image modals for posts its own file and exported function.

Renamed generatePostsUrl to paginationUrl as the name suggested it was in general when it was really only for pagination.

Removed unused handleSearch.js file

Other minor cleanups