New post display update
Branch with a new kind of post display without the old pagination. Might be easier regarding search/filter function without the pagination.

Added possibility to search/filter post ID as well as title, username and tags as before.

New file renderPosts.js now has the functions for displaying posts as well as a separate function for rendering the post content itself.

Removed now unused files pagination.js and fetchPosts.js.

Added JSDoc and comments to a few functions. Goal is to have JSDoc on most of or all functions and more comments for description.

Minor changes to style etc and cleanup.