Change images, JSDocs/comments
Added possibility of changing avatar image and banner image from a "settings" button on the profile page. The form inputs accepts the same kind of image URLs as the login, registration and post forms do.

Also added JSDocs to most/all functions as well as more comments throughout the code to hopefully make it easier to understand.