Neater post form
The post form is now "inside" an accordion.
Form requirements are displayed as small info icons the user can hover over to reveal input requirement on larger screens while on smallers screens theres the original requirements list, although beneath the form and not next to it.

Also added a post form to the profile page, however only on the logged-in users page