Filter/sort and better message system
Removed searchResults.html/.js/.scss alltogether and instead added the search bar and functionality to the feed.html page.

Added a select element users can use to pick between two options, "all posts" and "following" to sort posts if they want to see all or just from users they follow.

Added a .message-fixed element to all pages that works as a fixed message that is centered on the screen. Messages began to be a bit cluttered with too many message containers.